Move category subpage functions to separate file
